test(headless): split prompt A/B test boundaries
